requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
google-cloud-firestore
python-dotenv
pydantic

